% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmLogL.R
\name{tmLogL}
\alias{tmLogL}
\title{Log Likelihood Function for Text Mining}
\usage{
tmLogL(df, group, group.of.interest, text, sparsity = 0.8, threshold = 3)
}
\arguments{
\item{group}{Column name in your dataframe to group the text by}

\item{group.of.interest}{Group to compare to the rest}

\item{text}{Column name in your dataframe that contains the text}

\item{sparsity}{Level of acceptable sparsity in the text, default 0.80}

\item{threshold}{Chi-square statistic level at which to keep terms, default 3}

\item{dataframe}{Required data.frame containing the grouping varible and text}
}
\description{
This function allows you to compare the log 
likelihood of words between groups. To use, you must
supply the grouping variable, text variable and group of interest.
The function compares the expected frequency of words in the group of interest
and remaining text to the observed frequency. Words with high log likelihood
"surprisingly" occur more in the group of interest, while words with negative 
log likelihood occur less often than expected in the group of interest.
}
\examples{
tmLogL(df=my.data, group="City", group.of.interest="Chicago", text="Tweets")
tmLogL(df=other.data, group="Speaker", group.of.interest="George", text="Script.Lines", threshold=10)
}
\keyword{likelihood,}
\keyword{log}
\keyword{mining}
\keyword{text}
